Starting to implement the actual BMLT meeting parser.
The testing structure was changed, so I can easily inject a response dataset into the transport class.

I decided to make everything (except the meetings and formats) into classes, because there's a messy amount of interreferencing. I could probably structure it better, but this works great, and will be opaque to the consumer.